implement rqlite service management in BaseNeuron and update Validator to await service start
